Changes to use core in another project
* Core doesn't need curses, don't try to include it using our funky
  finder
* Ensure we don't define min/max since our wwiv cmake module isn't
  getting picked up when just adding core, so make that stand alonw
* Define SOCKET properly when compiling for x64.
* try to include the rapidjson bits from cereal using the existing
  include directory for it, not via deps.

+++ core/command_line.cpp
@@ -339,10 +339,10 @@ std::string CommandLineCommand::GetHelp() const {
   ss << program_name << " arguments:" << std::endl;
   auto max_len = 8;
   for (const auto& [_, c] : args_allowed_) {
-    max_len = std::max<int>(max_len, c.name_.size());
+    max_len = std::max<size_t>(max_len, c.name_.size());
   }
   for (const auto& a : commands_allowed_) {
-    max_len = std::max<int>(max_len, a.second->name().size());
+    max_len = std::max<size_t>(max_len, a.second->name().size());
   }
 
   for (const auto& [_, c] : args_allowed_) {
